+++ CoreAnalyticView/src/au/gov/asd/tac/constellation/views/analyticview/results/GraphResult.java
@@ -17,7 +17,7 @@ package au.gov.asd.tac.constellation.views.analyticview.results;
 
 import au.gov.asd.tac.constellation.graph.GraphElementType;
 import au.gov.asd.tac.constellation.views.analyticview.results.GraphResult.GraphScore;
-import java.util.List;
+import java.util.Collection;
 
 /**
  * Stores {@link GraphScore} objects which represent a float score which applies
@@ -34,7 +34,7 @@ public class GraphResult extends AnalyticResult<GraphScore> {
     }
 
     @Override
-    public void setSelectionOnVisualisation(final GraphElementType elementType, final List<Integer> elementIds) {
+    public void setSelectionOnVisualisation(final GraphElementType elementType, final Collection<Integer> elementIds) {
         // Overridden method intentionally left blank
     }