+++ CoreTimelineView/src/au/gov/asd/tac/constellation/views/timeline/clustering/ClusteringManager.java
@@ -211,7 +211,7 @@ public class ClusteringManager {
     }
 
     public Set<TreeElement> getElementsToDraw() {
-        return elementsToDraw;
+        return Collections.unmodifiableSet(elementsToDraw);
     }
 
     public void clearTree() {