+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Constellation Changes
 
+## Changes in April 2024
+-   Removed `graph` parameter from `PermanentMergeTableModel.initialise` as it was unused.
+
 ## Changes in March 2024
 -   Removed `dateTimeAttr` parameter from `ClusteringManager.InitDimOrHidePlugin` as it was unused.
 -   Removed `datetimeAtrr` parameter from `TimelinePanel.initExclusionState` as it was unused.